AppShell and Start up file
This commit holds the AppShell file which has the navigation menu and the MauiProgram is the start up file. The important thing in the start up file is the calling of the ConfigureSyncfusionCore() function which allows me to show the scheduler and register appointments! The argument in the function is the license key responsible to dont have the trial version of the scheduler!